Fix Bug In State copy constructor.
The output of pendulum_example.cpp showed that when a range of actions
was computed numerically, the output was the cumulative sum of the
actions and not the actions themselves.

This is due to the construction of a State2_Action from a State2 not taking
the .J() to zero.

Strangely, this was not true in every built. For example, when we std::cout
the intermediate results for debugging, the output was correct. Probably
having to do with optimizations we cannot easily control.